+++ memberaudit/managers.py
@@ -27,7 +27,11 @@ from allianceauth.authentication.models import CharacterOwnership
 from allianceauth.services.hooks import get_extension_logger
 
 from . import __title__
-from .constants import EVE_TYPE_ID_SOLAR_SYSTEM, EVE_CATEGORY_ID_SKILL
+from .constants import (
+    EVE_TYPE_ID_SOLAR_SYSTEM,
+    EVE_CATEGORY_ID_SKILL,
+    EVE_CATEGORY_ID_SHIP,
+)
 from .app_settings import MEMBERAUDIT_LOCATION_STALE_HOURS
 from .helpers import fetch_esi_status
 from .providers import esi
@@ -39,12 +43,24 @@ logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 BULK_METHODS_BATCH_SIZE = 500
 
 
+class EveShipTypeManger(models.Manager):
+    def get_queryset(self):
+        return (
+            super()
+            .get_queryset()
+            .select_related("eve_group")
+            .filter(published=True)
+            .filter(eve_group__eve_category_id=EVE_CATEGORY_ID_SHIP)
+        )
+
+
 class EveSkillTypeManger(models.Manager):
     def get_queryset(self):
         return (
             super()
             .get_queryset()
             .select_related("eve_group")
+            .filter(published=True)
             .filter(eve_group__eve_category_id=EVE_CATEGORY_ID_SKILL)
         )